functional id and range
