+++ smartampel/client/app/src/main/java/xlight/xlight/interfaces/UIUpdater.java
@@ -1,0 +1,11 @@
+package xlight.xlight.interfaces;
+
+/**
+ * Created by paul on 19.09.17.
+ */
+
+public interface UIUpdater {
+
+    void updateUi(String name, double distance);
+
+}
