+++ smartampel/client/app/src/main/res/layout/fragment_main.xml
@@ -14,11 +14,18 @@
         tools:layout_editor_absoluteY="8dp"
         tools:layout_editor_absoluteX="8dp">
 
-        <ListView
-            android:id="@+id/bluetooth_list"
+        <TextView
+            android:id="@+id/ibeacon_name_textview"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
-        </ListView>
+            android:layout_height="wrap_content"
+            android:text="iBeacon Name"
+            />
+        <TextView
+            android:id="@+id/ibeacon_range_textview"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@+id/ibeacon_name_textview"
+            android:layout_alignParentStart="true" />
 
     </RelativeLayout>
 </android.support.constraint.ConstraintLayout>
